Do not like about the types.

+++ engine.go
@@ -45,13 +45,13 @@ func (g *Game) Free() {
 	C.futhark_context_free(g.ctx)
 }
 
-func (g *Game) AddSphere(x, y, z, radius float64, colour uint32, shine float64) {
+func (g *Game) AddSphere(x, y, z, radius float32, colour uint32, shine float32) {
 	defer C.futhark_free_opaque_world(g.ctx, g.world)
 	var i C.int32_t
 	C.futhark_entry_add_sphere(g.ctx, &g.world, &i, g.world, C.float(x), C.float(y), C.float(z), C.float(radius), C.int32_t(colour), C.float(shine))
 }
 
-func (g *Game) SetSphereRadius(i int32, radius float64) {
+func (g *Game) SetSphereRadius(i int32, radius float32) {
 	defer C.futhark_free_opaque_world(g.ctx, g.world)
 	C.futhark_entry_set_sphere_radius(g.ctx, &g.world, g.world, C.int32_t(i), C.float(radius))
 }
@@ -68,7 +68,7 @@ func (g *Game) SetSpherePositions(xs []float32, ys []float32, zs []float32) {
 		g.world, xs_fut, ys_fut, zs_fut)
 }
 
-func (g *Game) AddPlane(pos_x, pos_y, pos_z, norm_x, norm_y, norm_z float64, colour uint32, shine float64) {
+func (g *Game) AddPlane(pos_x, pos_y, pos_z, norm_x, norm_y, norm_z float32, colour uint32, shine float32) {
 	defer C.futhark_free_opaque_world(g.ctx, g.world)
 	var i C.int32_t
 	C.futhark_entry_add_plane(g.ctx, &g.world, &i,
@@ -78,14 +78,14 @@ func (g *Game) AddPlane(pos_x, pos_y, pos_z, norm_x, norm_y, norm_z float64, col
 		C.int32_t(colour), C.float(shine))
 }
 
-func (g *Game) AddLight(x, y, z float64, colour uint32, intensity float64) {
+func (g *Game) AddLight(x, y, z float32, colour uint32, intensity float32) {
 	defer C.futhark_free_opaque_world(g.ctx, g.world)
 	C.futhark_entry_add_light(g.ctx, &g.world, g.world,
 		C.float(x), C.float(y), C.float(z), C.int32_t(colour),
 		C.float(intensity))
 }
 
-func (g *Game) Render(fov int, eye_pos_x, eye_pos_y, eye_pos_z, eye_dir_a, eye_dir_b float64, ambient uint32, ambient_intensity float64, limit int) {
+func (g *Game) Render(fov int, eye_pos_x, eye_pos_y, eye_pos_z, eye_dir_a, eye_dir_b float32, ambient uint32, ambient_intensity float32, limit int) {
 	var frame_fut *C.struct_futhark_i32_2d
 	C.futhark_entry_render(g.ctx, &frame_fut, g.world,
 		g.screenX, g.screenY, C.int32_t(fov),
